+++ src/Report.php
@@ -83,13 +83,13 @@ class Report
             Route::get('report/queue', 'QueuedReportController@index')->name('queue.index');
             Route::post('report/queue/{report}', 'QueuedReportController@render')->name('queue.render');
             Route::get('report/queue/job/{job}', 'QueuedReportController@job')->name('queue.job');
-            // TODO Route::get('report/queue/result/{job}', 'QueuedReportController@result')->name('queue.result');
             Route::get('report/queue/export/{job}', 'QueuedReportController@export')->name('queue.export');
 
             // Synchronous
             Route::get('report', 'ReportController@index')->name('index');
             Route::get('report/{report}', 'ReportController@show')->name('show');
             Route::post('report/{report}', 'ReportController@render')->name('render');
+            Route::post('report/{report}/preview', 'ReportController@preview')->name('preview');
 
             Route::post('report/{report}/export', 'ReportController@generateExport')->name('export');
         });